requirements.txt
asgiref==3.5.2
Django==3.2.15
django-debug-toolbar==3.7.0
django-environ==0.9.0
django-phonenumber-field==7.0.0
et-xmlfile==1.1.0
jdcal==1.4.1
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.2
pdfkit==1.0.0
phonenumbers==8.12.57
Pillow==9.2.0
psycopg2-binary==2.9.4
python-dateutil==2.8.2
pytz==2022.4
six==1.16.0
sqlparse==0.4.3
wkhtmltopdf==0.2
